SingleVendor/requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.4
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.12.0
django-rest-authtoken==2.1.3
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.2.0
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
